backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
groq>=0.9.0
networkx==3.3
pdfplumber==0.11.0
pypdf==4.2.0
python-docx==1.1.0
pydantic==2.7.1
python-multipart==0.0.9
python-dotenv==1.0.1
# NLP extraction layers (added v3.0)
spacy>=3.7.0
# Run after install: python -m spacy download en_core_web_sm
transformers>=4.40.0
torch>=2.0.0
